+++ zrml/combo/src/types/cryptographic_id_manager/typedefs.rs
@@ -1,0 +1,1 @@
+pub type Hash = [u8; 32];
